changed the sensitivity from int to float
similar reason to prior commit to give flexibility to us and the user

+++ Dev2-Prototype1/Assets/Scripts/cameraController.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class cameraController : MonoBehaviour
 {
-    [SerializeField] int sens;
+    [SerializeField] float sens;
     [SerializeField] int lockVertMin, lockVertMax; // Stops from looking too far up/down and inverting the camera
     [SerializeField] bool invertY;
     [SerializeField] Transform player;
